+++ src/www/defaults.ts
@@ -16,3 +16,37 @@ export const DEFAULT_OPEN_OPTIONS: HiddenInAppBrowserOpenOptions = {
   hideurlbar: true,
   fullscreen: true,
 };
+
+export const DEFAULT_EXTERNAL_BROWSER_OPTIONS: HiddenInAppBrowserOpenOptions = {
+  url: "",
+  hidden: false,
+  location: true,
+  toolbar: true,
+  zoom: true,
+  hardwareback: true,
+  mediaPlaybackRequiresUserAction: false,
+  shouldPauseOnSuspend: false,
+  clearsessioncache: false,
+  cache: true,
+  disallowoverscroll: false,
+  hidenavigationbuttons: false,
+  hideurlbar: false,
+  fullscreen: false,
+};
+
+export const DEFAULT_WEBVIEW_OPTIONS: HiddenInAppBrowserOpenOptions = {
+  url: "",
+  hidden: false,
+  location: true,
+  toolbar: true,
+  zoom: true,
+  hardwareback: true,
+  mediaPlaybackRequiresUserAction: false,
+  shouldPauseOnSuspend: false,
+  clearsessioncache: false,
+  cache: true,
+  disallowoverscroll: false,
+  hidenavigationbuttons: false,
+  hideurlbar: false,
+  fullscreen: false,
+};