+++ src/www/definitions.ts
@@ -23,4 +23,20 @@ export interface HiddenInAppBrowser {
     onSuccess?: () => void,
     onError?: (error: any) => void
   ): Promise<void>;
+  
+  openInExternalBrowser(
+    urlOrOptions: string | HiddenInAppBrowserOpenOptions,
+    target?: string,
+    optionsString?: string,
+    onSuccess?: () => void,
+    onError?: (error: any) => void
+  ): Promise<void>;
+  
+  openInWebView(
+    urlOrOptions: string | HiddenInAppBrowserOpenOptions,
+    target?: string,
+    optionsString?: string,
+    onSuccess?: () => void,
+    onError?: (error: any) => void
+  ): Promise<void>;
 }